pyproject.toml
[tool.poetry]
name = "01-simple-llm-app"
version = "0.1.0"
description = ""
authors = [" Riyadh <riyadhgenai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
python-dotenv = "^1.0.1"
langchain = "^0.2.11"
langchain-openai = "^0.1.20"
langserve = "^0.2.2"
jupyterlab = "^4.2.4"
fastapi = "^0.111.1"
sse-starlette = "^2.1.3"
langchain-groq = "0.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
